feat: Add code to print a square of size n
The code changes add functionality to print a square of size n, where n is a natural number input by the user. Each side of the square is made up of the character 'X'. This change improves the program by allowing the user to specify the size of the square to be printed.