feat: Add code to print a right triangle of size n
The code changes add functionality to print a right triangle of size n, where n is a natural number input by the user. Each line of the triangle is made up of the character 'X'. This change improves the program by allowing the user to specify the size of the triangle to be printed.