feat: Add code to sum natural numbers
The code changes add functionality to read a series of natural numbers from input and calculate their sum. The first number read determines the number of values to be summed. If the first number is non-negative, it represents the count of values to be read and summed. If the first number is -1, it indicates that the values will be read until the character "F" is encountered. This change improves the program by allowing the user to input a variable number of values to be summed.