Add function for finding prime numbers
The code changes in `Exercice-UpyLab-4.3.py` add a new function `premier(n)` that checks if a given positive integer `n` is a prime number. Additionally, the code includes a program that reads an integer `x` and prints all prime numbers less than `x` using the `premier` function. This enhancement improves the functionality of the code by providing a way to identify prime numbers.