feat: Add default values to somme function
The code changes in `Exercice-UpyLab-4.6.py` add default values to the `somme` function. The function now accepts two integer parameters `a` and `b`, with default values of 0 and 1 respectively. This enhancement improves the flexibility of the function by allowing it to be called with or without arguments, providing default values when necessary.