Write a function create_map(size, trapsNbr) which receives two integers as parameters, size, between 2 and 100, and trapsNbr, with a value strictly less than size x size, and which returns a dictionary implementing a map as in the previous example of size size and in which there are trapsNbr boxes containing a trap (modeled by the value -1) and a box containing a treasure (modeled by the value 1). The location of these boxes will be random.
Write a play_game(map_size, treasure_map) function that receives an integer and a map of size map_size x map_size, such as those obtained using the create_map function, and that asks the user to enter the coordinates of a square, up to 'to come across an occupied square. If the user found the treasure, the function returns True, otherwise the user ran into a trap and the function returns False.